requirements.txt
# KaliGPT Python Dependencies

# Core dependencies
pexpect>=4.8.0
requests>=2.31.0
rich>=13.0.0

# AI/LLM dependencies
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.25.0

# Report generation
markdown>=3.5.0
weasyprint>=60.0

# GUI dependencies
pillow>=10.0.0              # Desktop GUI - image support
flask>=3.0.0                # Web GUI - web framework
flask-socketio>=5.3.0       # Web GUI - WebSocket support
python-socketio>=5.10.0     # Web GUI - Socket.IO client
eventlet>=0.33.0            # Web GUI - async server

# Optional dependencies
# Uncomment if needed
# pdfkit>=1.0.0
# reportlab>=4.0.0
# jinja2>=3.1.0
